feat: Refactor and expand Module 1 lessons 1.1-1.5
